requirements.txt
customtkinter==5.2.2
requests
Pillow
openpyxl==3.1.2
pandas==2.2.2
pywin32==310
sentence-transformers==2.7.0
python-dotenv==1.0.0
google-generativeai>=0.3.0 